Set maximum size for compressed tarball scanning
This limit has been introduced due to the scanning for info of a compressed tarball (tar.gz/tar.bz2) inside the enveloping archive taking as much time as the extraction itself, because in practice the archive must be uncompressed before the contents of the tarball can be scanned for info.

The bigger the archive the less practical it becomes to scan its contents, performance-wise. An input box for this limit has been created in Settings which allows the user to set the desired value, after which no scanning takes place and the only info presented is the assumed tarball inside the archive.